+++ api/jobs_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/hashicorp/nomad/api/testutil"
+	"github.com/hashicorp/nomad/api/internal/testutil"
 	"github.com/kr/pretty"
 	"github.com/stretchr/testify/require"
 )
