+++ internal/testing/apitests/structsync_test.go
